// server/lib/hxTorquescript.d.ts
export class expr_BreakStmt extends expr_Stmt {
	constructor(lineNo: number);
	print(indent: number, isStmt: boolean): string;
	precompileStmt(compiler: Compiler, loopCount: number): number;
	compileStmt(compiler: Compiler, context: CompileContext): number;
	visitStmt(optimizerPass: IASTVisitor): void;
}

export class expr_ContinueStmt extends expr_Stmt {
	constructor(lineNo: number);
	print(indent: number, isStmt: boolean): string;
	precompileStmt(compiler: Compiler, loopCount: number): number;
	compileStmt(compiler: Compiler, context: CompileContext): number;
	visitStmt(optimizerPass: IASTVisitor): void;
}

export class expr_Expr extends expr_Stmt {
	constructor(lineNo: number);
	_hx_constructor(lineNo: number): void;
	print(indent: number, isStmt: boolean): string;
	precompileStmt(compiler: Compiler, loopCount: number): number;
	compileStmt(compiler: Compiler, context: CompileContext): number;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(optimizerPass: IASTVisitor): void;
	static conversionOp(src: any, dest: any): expr_OpCode;
}

export class expr_ParenthesisExpr extends expr_Expr {
	constructor(expr: expr_Expr);
	expr: expr_Expr;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(optimizerPass: IASTVisitor): void;
}

export class expr_ReturnStmt extends expr_Stmt {
	constructor(lineNo: number, expr: expr_Expr);
	expr: expr_Expr;
	print(indent: number, isStmt: boolean): string;
	precompileStmt(compiler: Compiler, loopCount: number): number;
	compileStmt(compiler: Compiler, context: CompileContext): number;
	visitStmt(optimizerPass: IASTVisitor): void;
}

export class expr_IfStmt extends expr_Stmt {
	constructor(lineNo: number, condition: expr_Expr, body: expr_Stmt[], elseBlock: expr_Stmt[]);
	condition: expr_Expr;
	body: expr_Stmt[];
	elseBlock: expr_Stmt[];
	integer: boolean;
	endifOffset: number;
	elseOffset: number;
	ifStmtList: boolean;
	elseStmtList: boolean;
	print(indent: number, isStmt: boolean): string;
	precompileStmt(compiler: Compiler, loopCount: number): number;
	compileStmt(compiler: Compiler, context: CompileContext): number;
	visitStmt(optimizerPass: IASTVisitor): void;
}

export class expr_LoopStmt extends expr_Stmt {
	constructor(lineNo: number, condition: expr_Expr, init: expr_Expr, end: expr_Expr, body: expr_Stmt[]);
	_hx_constructor(lineNo: number, condition: expr_Expr, init: expr_Expr, end: expr_Expr, body: expr_Stmt[]): void;
	condition: expr_Expr;
	init: expr_Expr;
	end: expr_Expr;
	body: expr_Stmt[];
	integer: boolean;
	loopBlockStartOffset: number;
	continueOffset: number;
	breakOffset: number;
	isForLoop: boolean;
	isStatementList: boolean;
	print(indent: number, isStmt: boolean): string;
	precompileStmt(compiler: Compiler, loopCount: number): number;
	compileStmt(compiler: Compiler, context: CompileContext): number;
	visitStmt(optimizerPass: IASTVisitor): void;
}

export class Token {
	constructor(type: any, lexeme: string, literal: any, line: number, position: number);
	type: any;
	lexeme: string;
	literal: any;
	line: number;
	position: number;
}

export class expr_BinaryExpr extends expr_Expr {
	constructor(lineNo: number);
	_hx_constructor(lineNo: number): void;
	left: expr_Expr;
	right: expr_Expr;
	op: Token;
	optimized: boolean;
	optimizedExpr: expr_Expr;
	print(indent: number, isStmt: boolean): string;
}

export class expr_FloatBinaryExpr extends expr_BinaryExpr {
	constructor(left: expr_Expr, right: expr_Expr, op: Token);
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	visitStmt(visitor: IASTVisitor): void;
	getPrefferredType(): any;
}

export class expr_IntBinaryExpr extends expr_BinaryExpr {
	constructor(left: expr_Expr, right: expr_Expr, op: Token);
	subType: any;
	operand: expr_OpCode;
	getSubTypeOperand(): void;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	visitStmt(visitor: IASTVisitor): void;
	getPrefferredType(): any;
}

export class expr_StrEqExpr extends expr_BinaryExpr {
	constructor(left: expr_Expr, right: expr_Expr, op: Token);
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	visitStmt(visitor: IASTVisitor): void;
	getPrefferredType(): any;
}

export class expr_StrCatExpr extends expr_BinaryExpr {
	constructor(left: expr_Expr, right: expr_Expr, op: Token);
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	visitStmt(visitor: IASTVisitor): void;
	getPrefferredType(): any;
}

export class expr_CommaCatExpr extends expr_BinaryExpr {
	constructor(left: expr_Expr, right: expr_Expr);
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
}

export class expr_ConditionalExpr extends expr_Expr {
	constructor(condition: expr_Expr, trueExpr: expr_Expr, falseExpr: expr_Expr);
	condition: expr_Expr;
	trueExpr: expr_Expr;
	falseExpr: expr_Expr;
	integer: boolean;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(visitor: IASTVisitor): void;
}

export class expr_IntUnaryExpr extends expr_Expr {
	constructor(expr: expr_Expr, op: Token);
	_hx_constructor(expr: expr_Expr, op: Token): void;
	expr: expr_Expr;
	op: Token;
	integer: boolean;
	optimized: boolean;
	optimizedExpr: expr_Expr;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(visitor: IASTVisitor): void;
}

export class expr_FloatUnaryExpr extends expr_Expr {
	constructor(expr: expr_Expr, op: Token);
	_hx_constructor(expr: expr_Expr, op: Token): void;
	expr: expr_Expr;
	op: Token;
	optimized: boolean;
	optimizedExpr: expr_Expr;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(visitor: IASTVisitor): void;
}
// ...
export class expr_IntExpr extends expr_Expr {
	constructor(lineNo: number, value: number);
	value: number;
	index: number;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(visitor: IASTVisitor): void;
}

export class expr_FloatExpr extends expr_Expr {
	constructor(lineNo: number, value: number);
	value: number;
	index: number;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(visitor: IASTVisitor): void;
}
// ...
export class expr_ConstantExpr extends expr_Expr {
	constructor(name: Token);
	name: Token;
	fVal: number;
	index: number;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(visitor: IASTVisitor): void;
}

export class expr_AssignExpr extends expr_Expr {
	constructor(varExpr: expr_VarExpr, expr: expr_Expr);
	varExpr: expr_VarExpr;
	expr: expr_Expr;
	subType: any;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(visitor: IASTVisitor): void;
}

export class expr_AssignOpExpr extends expr_Expr {
	constructor(varExpr: expr_VarExpr, expr: expr_Expr, op: Token);
	varExpr: expr_VarExpr;
	expr: expr_Expr;
	op: Token;
	subType: any;
	operand: expr_OpCode;
	getAssignOpTypeOp(): void;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(visitor: IASTVisitor): void;
}

export const enum expr_FuncCallType {
	FunctionCall = 0,
	MethodCall = 1,
	ParentCall = 2,
}

export class expr_FuncCallExpr extends expr_Expr {
	constructor(name: Token, namespace: Token, args: expr_Expr[], callType: expr_FuncCallType);
	name: Token;
	namespace: Token;
	args: expr_Expr[];
	callType: expr_FuncCallType;
	print(indent: number, isStmt: boolean): string;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(optimizerPass: IASTVisitor): void;
}
// ...
export class expr_ObjectDeclExpr extends expr_Expr {
	constructor(className: expr_Expr, parentObject: Token, objectNameExpr: expr_Expr, args: expr_Expr[], slotDecls: expr_SlotAssignExpr[], subObjects: expr_ObjectDeclExpr[], structDecl: boolean);
	_hx_constructor(className: expr_Expr, parentObject: Token, objectNameExpr: expr_Expr, args: expr_Expr[], slotDecls: expr_SlotAssignExpr[], subObjects: expr_ObjectDeclExpr[], structDecl: boolean): void;
	className: expr_Expr;
	parentObject: Token;
	objectNameExpr: expr_Expr;
	args: expr_Expr[];
	slotDecls: expr_SlotAssignExpr[];
	subObjects: expr_ObjectDeclExpr[];
	structDecl: boolean;
	failOffset: number;
	print(indent: number, isStmt: boolean): string;
	precompileSubObject(compiler: Compiler, typeReq: any): number;
	compileSubObject(compiler: Compiler, context: CompileContext, typeReq: any, root: boolean): number;
	precompile(compiler: Compiler, typeReq: any): number;
	compile(compiler: Compiler, context: CompileContext, typeReq: any): number;
	getPrefferredType(): any;
	visitStmt(optimizerPass: IASTVisitor): void;
}

export class expr_FunctionDeclStmt extends expr_Stmt {
	constructor(functionName: Token, args: expr_VarExpr[], stmts: expr_Stmt[], namespace: Token);
	functionName: Token;
	packageName: Token;
	args: expr_VarExpr[];
	stmts: expr_Stmt[];
	namespace: Token;
	endOffset: number;
	argc: number;
	print(indent: number, isStmt: boolean): string;
	precompileStmt(compiler: Compiler, loopCount: number): number;
	compileStmt(compiler: Compiler, context: CompileContext): number;
	visitStmt(optimizerPass: IASTVisitor): void;
}

export interface IASTVisitor {
	visitStmt(stmt: expr_Stmt): void;
	visitBreakStmt(stmt: expr_BreakStmt): void;
	visitContinueStmt(stmt: expr_ContinueStmt): void;
	visitExpr(expr: expr_Expr): void;
	visitParenthesisExpr(expr: expr_ParenthesisExpr): void;
	visitReturnStmt(stmt: expr_ReturnStmt): void;
	visitIfStmt(stmt: expr_IfStmt): void;
	visitLoopStmt(stmt: expr_LoopStmt): void;
	visitBinaryExpr(expr: expr_BinaryExpr): void;
	visitFloatBinaryExpr(expr: expr_FloatBinaryExpr): void;
	visitIntBinaryExpr(expr: expr_IntBinaryExpr): void;
	visitStrEqExpr(expr: expr_StrEqExpr): void;
	visitStrCatExpr(expr: expr_StrCatExpr): void;
	visitCommatCatExpr(expr: expr_CommaCatExpr): void;
	visitConditionalExpr(expr: expr_ConditionalExpr): void;
	visitIntUnaryExpr(expr: expr_IntUnaryExpr): void;
	visitFloatUnaryExpr(expr: expr_FloatUnaryExpr): void;
	visitVarExpr(expr: expr_VarExpr): void;
	visitIntExpr(expr: expr_IntExpr): void;
	visitFloatExpr(expr: expr_FloatExpr): void;
	visitStringConstExpr(expr: expr_StringConstExpr): void;
	visitConstantExpr(expr: expr_ConstantExpr): void;
	visitAssignExpr(expr: expr_AssignExpr): void;
	visitAssignOpExpr(expr: expr_AssignOpExpr): void;
	visitFuncCallExpr(expr: expr_FuncCallExpr): void;
	visitSlotAccessExpr(expr: expr_SlotAccessExpr): void;
	visitSlotAssignExpr(expr: expr_SlotAssignExpr): void;
	visitSlotAssignOpExpr(expr: expr_SlotAssignOpExpr): void;
	visitObjectDeclExpr(expr: expr_ObjectDeclExpr): void;
	visitFunctionDeclStmt(stmt: expr_FunctionDeclStmt): void;
}
// ...
export class Parser {
	constructor(tokens: Token[]);
	parse(): expr_Stmt[];
}

export class Scanner {
	constructor(s: string);
	scanTokens(): Token[];
	static unescape(s: string): string;
	static escape(s: string): string;
}